// frontend/src/core/api/generated/model/dayRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { DayRecordReason } from './dayRecordReason'
import type { DayRecordStatus } from './dayRecordStatus'

export interface DayRecord {
  status: DayRecordStatus
  /** @nullable */
  reason?: DayRecordReason
  /** @nullable */
  checkedInAt?: string | null
  /** @nullable */
  siteName?: string | null
  /** @nullable */
  markedByName?: string | null
  /** @nullable */
  markedAt?: string | null
}