// frontend/src/core/api/generated/model/markMemberRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { MarkMemberRequestReason } from './markMemberRequestReason'
import type { MarkMemberRequestStatus } from './markMemberRequestStatus'

export interface MarkMemberRequest {
  status: MarkMemberRequestStatus
  /** @nullable */
  reason?: MarkMemberRequestReason
}