// frontend/src/core/api/generated/model/markMemberRequestReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

/**
 * @nullable
 */
export type MarkMemberRequestReason =
  (typeof MarkMemberRequestReason)[keyof typeof MarkMemberRequestReason] | null

export const MarkMemberRequestReason = {
  SICK: 'SICK',
  UNEXCUSED: 'UNEXCUSED',
  OTHER: 'OTHER',
} as const